requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
Django==2.1.2
djangorestframework==3.8.2
Faker==0.9.1
gunicorn==19.9.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
kombu==4.2.1
Markdown==3.0.1
MarkupSafe==1.0
mixer==6.1.3
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.5
redis==2.10.6
requests==2.19.1
six==1.11.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
